14/12/2015 003
Cambio para la precisión del sistema

+++ functions/US_FRETORNA_ID_USUARIO.sql
@@ -3,7 +3,7 @@
 CREATE OR REPLACE FUNCTION US_FRETORNA_ID_USUARIO (    p_usuario        VARCHAR(50) ) RETURNS INTEGER AS $$
       DECLARE 
       
-         v_tius         integer := 0;         
+         v_tius         bigint := 0;         
          
          c_usuario CURSOR FOR
          SELECT tius_tius
