+++ functions/US_FVERIFICA_CORREO.sql
@@ -3,7 +3,7 @@ CREATE FUNCTION US_FVERIFICA_CORREO(
                                     ) RETURNS VARCHAR AS'
    DECLARE
    
-   v_num_correo    int := 0;
+   v_num_correo    bigint := 0;
    
    c_correo CURSOR FOR 
       SELECT count(*)
