+++ functions/US_FVERIFICA_USUARIO.sql
@@ -3,7 +3,7 @@ CREATE FUNCTION US_FVERIFICA_USUARIO(
                                     ) RETURNS VARCHAR AS'
    DECLARE
    
-   v_num_usuario    int := 0;
+   v_num_usuario    bigint := 0;
    
    c_usuario CURSOR FOR 
       SELECT count(*)
